+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/batch/BatchOrderInvoicingSale.java
@@ -24,6 +24,8 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import com.axelor.apps.account.db.repo.InvoiceRepository;
+import com.axelor.apps.base.db.repo.BlockingRepository;
+import com.axelor.apps.base.service.BlockingService;
 import com.axelor.apps.sale.db.SaleOrder;
 import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.apps.supplychain.db.SupplychainBatch;
@@ -85,6 +87,9 @@ public class BatchOrderInvoicingSale extends BatchOrderInvoicing {
 				+ "AND (invoice.saleOrder = self "
 				+ "OR invoice.saleOrder IS NULL AND EXISTS (SELECT 1 FROM invoice.invoiceLineList invoiceLine "
 				+ "WHERE invoiceLine.saleOrderLine MEMBER OF self.saleOrderLineList)))");
+
+		filterList.add("self.partner.id NOT IN (" + Beans.get(BlockingService.class).listOfBlockedPartner(supplychainBatch.getCompany(), BlockingRepository.INVOICING_BLOCKING) + ")");
+
 		query.bind("invoiceStatusSelect", InvoiceRepository.STATUS_CANCELED);
 
 		List<Long> anomalyList = Lists.newArrayList(0L);